+++ marketing_event-20210101/ChangeLog.md
@@ -1,3 +1,10 @@
+2024-09-09 Version: 1.0.1
+- Update API FindGuestCredentialsRecord: add param EndDateTime.
+- Update API FindGuestCredentialsRecord: add param StartDateTime.
+- Update API FindGuestTicketRecord: add param EndDateTime.
+- Update API FindGuestTicketRecord: add param StartDateTime.
+
+
 2024-09-04 Version: 1.0.0
 - Generated python 2021-01-01 for marketing_event.